Switcheroo on util.py to Steve's version
