+++ data_hough_detect.py
@@ -79,7 +79,7 @@ def map_diff2(z,ndiff):
 
 
 def map_unravel(z,params):
-    import util2
+    import util
     new_maps =[]
     
     for wave in z:
